Add platform_specific_funcs
修正模块名称中的拼写错误；
添加数个新的平台差异功能函数；
