+++ src/py2exe_gui/Utilities/__init__.py
@@ -4,6 +4,6 @@
 """公共基础功能类与函数"""
 
 from .open_qfile import QtFileOpen
-from .platform_specifc_funcs import get_sys_python, open_dir_in_explorer
+from .platform_specific_funcs import *
 from .python_env import ALL_PY_ENVs, PyEnv
 from .qobject_tr import QObjTr
